# Conflicts:
#	tests/c2me-test/build.gradle

+++ c2me-rewrites-chunkio/src/main/java/com/ishland/c2me/rewrites/chunkio/common/C2MEStorageThread.java
@@ -32,10 +32,10 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.Executor;
 import java.util.concurrent.PriorityBlockingQueue;
-import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.LockSupport;
 import java.util.function.Function;
@@ -47,6 +47,7 @@ public class C2MEStorageThread extends Thread {
     private static final Logger LOGGER = LoggerFactory.getLogger("C2ME Storage");
 
     private static final AtomicLong SERIAL = new AtomicLong(0);
+    public static final int SEMAPHORE_COUNTER_INITIAL = (int) ModuleEntryPoint.globalExecutorParallelism + 1;
 
     private final AtomicBoolean closing = new AtomicBoolean(false);
     private final CompletableFuture<Void> closeFuture = new CompletableFuture<>();
@@ -56,7 +57,8 @@ public class C2MEStorageThread extends Thread {
     private final Long2ObjectLinkedOpenHashMap<NbtCompound> writeBacklog = new Long2ObjectLinkedOpenHashMap<>();
     private final Long2ObjectOpenHashMap<NbtCompound> cache = new Long2ObjectOpenHashMap<>();
     private final Queue<ReadRequest> pendingReadRequests;
-    private final Semaphore readSemaphore = new Semaphore((int) ModuleEntryPoint.globalExecutorParallelism + 1);
+    private final AtomicInteger readSemaphoreCounter = new AtomicInteger(SEMAPHORE_COUNTER_INITIAL);
+    private final AtomicInteger writeSemaphoreCounter = new AtomicInteger(SEMAPHORE_COUNTER_INITIAL);
     private final ConcurrentLinkedQueue<WriteRequest> pendingWriteRequests = new ConcurrentLinkedQueue<>();
     private final ConcurrentLinkedQueue<Runnable> pendingTasks = new ConcurrentLinkedQueue<>();
     private final Executor executor = command -> {
@@ -126,9 +128,9 @@ public class C2MEStorageThread extends Thread {
         }
         this.pendingReadRequests.add(new ReadRequest(pos, future, scanner, this.priorityProvider != null ? this.priorityProvider.apply(pos) : null));
         LockSupport.unpark(this);
-        future.orTimeout(60, TimeUnit.SECONDS).exceptionally(throwable -> {
+        future.thenApply(Function.identity()).orTimeout(60, TimeUnit.SECONDS).exceptionally(throwable -> {
             if (throwable instanceof TimeoutException) {
-                LOGGER.warn("Chunk read at pos {} took too long (> 1min)", new ChunkPos(pos).toLong());
+                LOGGER.warn("Chunk read at pos {} took too long (> 1min)", new ChunkPos(pos));
             }
             return null;
         });
@@ -154,6 +156,11 @@ public class C2MEStorageThread extends Thread {
                 if (handlePendingReads()) continue;
                 if (handlePendingWrites()) continue;
                 if (writeBacklog()) continue;
+
+                if (isBusy()) {
+                    LockSupport.parkNanos("Waiting for tasks", 10_000_000);
+                    continue;
+                }
 
                 break;
             }
@@ -184,6 +191,17 @@ public class C2MEStorageThread extends Thread {
         return hasWork;
     }
 
+    private boolean isBusy() {
+        return this.writeSemaphoreCounter.get() < SEMAPHORE_COUNTER_INITIAL ||
+                this.readSemaphoreCounter.get() < SEMAPHORE_COUNTER_INITIAL ||
+                !this.pendingWriteRequests.isEmpty() ||
+                !this.pendingReadRequests.isEmpty() ||
+                !this.writeBacklog.isEmpty() ||
+                !this.pendingTasks.isEmpty() ||
+                !this.writeFutures.isEmpty() ||
+                !this.cache.isEmpty();
+    }
+
     private boolean handlePendingWrites() {
         boolean hasWork = false;
         WriteRequest writeRequest;
@@ -197,7 +215,7 @@ public class C2MEStorageThread extends Thread {
 
     private boolean handlePendingReads() {
         boolean hasWork = false;
-        while (!pendingReadRequests.isEmpty() && readSemaphore.tryAcquire()) {
+        while (!pendingReadRequests.isEmpty() && readSemaphoreCounter.get() > 0) {
             ReadRequest readRequest = this.pendingReadRequests.poll();
             hasWork = true;
             assert readRequest != null;
@@ -207,7 +225,6 @@ public class C2MEStorageThread extends Thread {
             final NbtCompound cached = this.cache.get(pos);
             if (cached != null) {
                 future.complete(cached);
-                readSemaphore.release();
                 continue;
             }
             scheduleChunkRead(pos, future, scanner);
@@ -216,7 +233,7 @@ public class C2MEStorageThread extends Thread {
     }
 
     private boolean writeBacklog() {
-        if (!this.writeBacklog.isEmpty()) {
+        if (!this.writeBacklog.isEmpty() && writeSemaphoreCounter.get() > 0) {
             final long pos = this.writeBacklog.firstLongKey();
             final NbtCompound nbt = this.writeBacklog.removeFirst();
             writeChunk(pos, nbt);
@@ -235,19 +252,18 @@ public class C2MEStorageThread extends Thread {
     private void doPriorityChanges() {
         if (this.pendingReadRequests instanceof PriorityBlockingQueue<ReadRequest> queue) {
             final long currentTimeMillis = System.currentTimeMillis();
-            if (currentTimeMillis > lastRebuild + 500) { // at most twice a second
+            final int currentPrioritySerial = PriorityUtils.priorityChangeSerial();
+            if (this.lastPrioritySerial != currentPrioritySerial &&
+                    (currentTimeMillis > lastRebuild + 500 || currentPrioritySerial - this.lastPrioritySerial >= 20)) { // rebuild only when every 500ms or 20 changes
                 lastRebuild = currentTimeMillis;
-                final int currentPrioritySerial = PriorityUtils.priorityChangeSerial();
-                if (this.lastPrioritySerial != currentPrioritySerial) {
-                    this.lastPrioritySerial = currentPrioritySerial;
-                    final long startTime = System.nanoTime();
-                    // re-add locks to reflect priority changes
-                    priorityChangeTmpStorage.clear();
-                    queue.drainTo(priorityChangeTmpStorage);
-                    queue.addAll(priorityChangeTmpStorage);
-                    priorityChangeTmpStorage.clear();
+                this.lastPrioritySerial = currentPrioritySerial;
+                final long startTime = System.nanoTime();
+                // re-add locks to reflect priority changes
+                priorityChangeTmpStorage.clear();
+                queue.drainTo(priorityChangeTmpStorage);
+                queue.addAll(priorityChangeTmpStorage);
+                priorityChangeTmpStorage.clear();
 //                System.out.printf("Did priority changes for %d entries in %.2fms\n", size, (System.nanoTime() - startTime) / 1_000_000.0);
-                }
             }
         }
     }
@@ -272,11 +288,9 @@ public class C2MEStorageThread extends Thread {
         if (cached != null) {
             if (scanner != null) {
                 cached.accept(scanner);
-                readSemaphore.release();
                 future.complete(null);
                 return;
             } else {
-                readSemaphore.release();
                 future.complete(cached);
                 return;
             }
@@ -288,9 +302,9 @@ public class C2MEStorageThread extends Thread {
             final DataInputStream chunkInputStream = regionFile.getChunkInputStream(pos1);
             if (chunkInputStream == null) {
                 future.complete(null);
-                readSemaphore.release();
                 return;
             }
+            readSemaphoreCounter.incrementAndGet();
             CompletableFuture.supplyAsync(() -> {
                 try {
                     try (DataInputStream inputStream = chunkInputStream) {
@@ -306,27 +320,30 @@ public class C2MEStorageThread extends Thread {
                     return null; // Unreachable anyway
                 }
             }, GlobalExecutors.executor).handle((compound, throwable) -> {
-                readSemaphore.release();
+                readSemaphoreCounter.decrementAndGet();
                 if (throwable != null) future.completeExceptionally(throwable);
                 else future.complete(compound);
                 return null;
             });
         } catch (Throwable t) {
-            readSemaphore.release();
             future.completeExceptionally(t);
         }
     }
 
     private void writeChunk(long pos, NbtCompound nbt) {
         if (nbt == null) {
-            try {
-                final ChunkPos pos1 = new ChunkPos(pos);
-                final RegionFile regionFile = ((IRegionBasedStorage) this.storage).invokeGetRegionFile(pos1);
-                regionFile.delete(pos1);
-            } catch (Throwable t) {
-                LOGGER.error("Error writing chunk %s".formatted(new ChunkPos(pos)), t);
+            if (this.cache.get(pos) == null) {
+                try {
+                    final ChunkPos pos1 = new ChunkPos(pos);
+                    final RegionFile regionFile = ((IRegionBasedStorage) this.storage).invokeGetRegionFile(pos1);
+                    regionFile.delete(pos1);
+                } catch (Throwable t) {
+                    LOGGER.error("Error writing chunk %s".formatted(new ChunkPos(pos)), t);
+                }
+                this.cache.remove(pos);
             }
         } else {
+            writeSemaphoreCounter.incrementAndGet();
             final CompletableFuture<Void> future = CompletableFuture.supplyAsync(() -> {
                 try {
                     final RawByteArrayOutputStream out = new RawByteArrayOutputStream(8096);
@@ -356,12 +373,13 @@ public class C2MEStorageThread extends Thread {
                     } catch (Throwable t) {
                         SneakyThrow.sneaky(t);
                     }
+                    this.cache.remove(pos);
                 }
             }, this.executor).handleAsync((unused, throwable) -> {
+                writeSemaphoreCounter.decrementAndGet();
                 if (throwable != null) LOGGER.error("Error writing chunk %s".formatted(new ChunkPos(pos)), throwable);
                 // TODO error retry
 
-                this.cache.remove(pos);
                 return null;
             }, this.executor);
             this.writeFutures.add(future);
